Changed float scale double scale in Picture

+++ app/src/main/java/com/example/rent/application1/models/Picture.java
@@ -53,7 +53,7 @@ public class Picture implements Parcelable{
         return "https://unsplash.it/300/200/?image=" + picture.getId();
     }
         //// TODO: 2017-08-22 scale
-    public static String getPhotoUrl(Picture picture, float scale) {
+    public static String getPhotoUrl(Picture picture, double scale) {
         return "https://unsplash.it/"+ picture.getWidth() * scale + "/" + picture.getWidth() * scale + "/?image=" + picture.getId();
     }
